+++ landlab/grid/hex.py
@@ -6,15 +6,12 @@ Do NOT add new documentation here. Grid documentation is now built in a
 semi- automated fashion. To modify the text seen on the web, edit the
 files `docs/text_for_[gridfile].py.txt`.
 """
-from functools import cached_property
-
 import numpy
 import xarray as xr
 
 from ..core.utils import as_id_array
 from ..graph import DualHexGraph
 from .base import ModelGrid
-from .linkorientation import LinkOrientation
 
 
 class HexModelGrid(DualHexGraph, ModelGrid):
@@ -251,145 +248,6 @@ class HexModelGrid(DualHexGraph, ModelGrid):
         half_nc = (nc + 1) // 2
         col = 2 * (n_mod_nc % half_nc) + n_mod_nc // half_nc
         return (row, col)
-
-    @cached_property
-    def orientation_of_link(self):
-        """Return array of link orientation codes (one value per link).
-
-        Orientation codes are defined by :class:`~.LinkOrientation`;
-        1 = E, 2 = ENE, 4 = NNE, 8 = N, 16 = NNW, 32 = ESE (using powers
-        of 2 allows for future applications that might want additive
-        combinations).
-
-        Examples
-        --------
-        >>> from landlab import HexModelGrid
-        >>> import numpy as np
-        >>> grid = HexModelGrid((3, 2))
-        >>> grid.orientation_of_link
-        array([ 1, 16,  4, 16,  4,  1,  1,  4, 16,  4, 16,  1])
-        >>> grid = HexModelGrid((2, 3), orientation="vertical")
-        >>> grid.orientation_of_link
-        array([32,  2,  8,  2, 32,  8,  8, 32,  2,  8,  2, 32])
-        """
-        code = numpy.round(self.angle_of_link * 6.0 / numpy.pi).astype(numpy.uint8)
-        code[code == 11] = 5
-        code[:] = 2**code
-
-        return code
-
-    @cached_property
-    def parallel_links_at_link(self):
-        """Return similarly oriented links connected to each link.
-
-        Return IDs of links of the same orientation that are connected to
-        each given link's tail or head node.
-
-        The data structure is a *numpy* array of shape ``(n_links, 2)`` containing the
-        IDs of the "tail-wise" (connected to tail node) and "head-wise" (connected
-        to head node) links, or -1 if the link is inactive (e.g., on the perimeter)
-        or it has no attached parallel neighbor in the given direction.
-
-        For instance, consider a 3x3 hex, in which link IDs are as shown::
-
-               o---17--o---18--o
-              / .     / .     / .
-             11 12   13  14  15  16
-            /     . /     . /     .
-           o---8---o---9---o---10--o
-            .     / .     / .     /
-             2   3   4   5   6   7
-              . /     . /     . /
-               o---0---o---1---o
-
-        Here's a mapping of the tail-wise and head-wise links, where
-        there are valid parallel links::
-
-               o-------o-------o
-              / .     / .     / .
-             /   .   /   .   /   .
-            /     4 3     6 5     .
-           o-----9-o-8--10-o-9-----o
-            .    13 12   15 14    /
-             .   /   .   /   .   /
-              . /     . /     . /
-               o-------o-------o
-
-        The corresponding data structure would be mostly filled with -1, but
-        for the 11 active links, it would look like::
-
-            3: [[-1, 13],
-            4:  [-1, 12],
-            5:  [-1, 15],
-            6:  [-1, 14],
-            8:  [-1,  9],
-            9:  [ 8, 10],
-            10: [ 9, -1],
-            12: [ 4, -1],
-            13: [ 3, -1],
-            14: [ 6, -1],
-            15: [ 5, -1]]
-
-        Examples
-        --------
-        >>> from landlab import HexModelGrid
-        >>> grid = HexModelGrid((3, 3))
-        >>> pll = grid.parallel_links_at_link
-        >>> pll[3:16]
-        array([[-1, 13],
-               [-1, 12],
-               [-1, 15],
-               [-1, 14],
-               [-1, -1],
-               [-1,  9],
-               [ 8, 10],
-               [ 9, -1],
-               [-1, -1],
-               [ 4, -1],
-               [ 3, -1],
-               [ 6, -1],
-               [ 5, -1]])
-        """
-        if self.orientation == "horizontal":
-            orientations = (LinkOrientation.E, LinkOrientation.NNE, LinkOrientation.NNW)
-        else:
-            orientations = (LinkOrientation.ENE, LinkOrientation.N, LinkOrientation.ESE)
-
-        links_at_node = self._oriented_links_at_node()
-
-        pll = numpy.full((self.number_of_links, 2), -1, dtype=int)
-
-        for col, orientation in enumerate(orientations):
-            links = self.orientation_of_link == orientation
-
-            if orientation == LinkOrientation.ESE:
-                pll[links, 0] = links_at_node[self.node_at_link_tail[links]][:, col]
-                pll[links, 1] = links_at_node[self.node_at_link_head[links]][:, col + 3]
-            else:
-                pll[links, 0] = links_at_node[self.node_at_link_tail[links]][:, col + 3]
-                pll[links, 1] = links_at_node[self.node_at_link_head[links]][:, col]
-
-        return pll
-
-    def _oriented_links_at_node(self):
-        if self.orientation == "horizontal":
-            orientations = (LinkOrientation.E, LinkOrientation.NNE, LinkOrientation.NNW)
-        else:
-            orientations = (LinkOrientation.ENE, LinkOrientation.N, LinkOrientation.ESE)
-
-        links_at_node = numpy.full((self.number_of_nodes, 6), -1, dtype=int)
-
-        for col, orientation in enumerate(orientations):
-            links = numpy.where(self.orientation_of_link == orientation)[0]
-
-            if orientation == LinkOrientation.ESE:
-                links_at_node[self.node_at_link_tail[links], col + 3] = links
-                links_at_node[self.node_at_link_head[links], col] = links
-            else:
-                links_at_node[self.node_at_link_tail[links], col] = links
-                links_at_node[self.node_at_link_head[links], col + 3] = links
-
-        return links_at_node
 
     def _configure_hexplot(self, data, data_label=None, color_map=None):
         """Sets up necessary information for making plots of the hexagonal grid
